Remove the title from the index file
If the title variable is set in the index file, the header area will
become cluttered.

Additionally if the title variable is set in both _config.yml and index
files, the one in the index takes precedence.

For title preservation and clean page layout, do not set the title in
the index.